[~] Design changes [~] Bugfixes [+] EngDub-GerSub Language Flag Icon added

+++ AniWorldReminder_Webpanel/Models/AniListAPIQuery.cs
@@ -78,7 +78,7 @@
         public bool? IsAdult { get; set; }
 
         [JsonProperty("averageScore")]
-        public int? AverageScore { get; set; }
+        public decimal? AverageScore { get; set; }
 
         [JsonProperty("nextAiringEpisode")]
         public object NextAiringEpisode { get; set; }
